refactor: extract TerminalContent — shared rendering for both modes
Resolves all deferred hickey/lowy findings except L2 (localStorage→server):

- C1: Extract TerminalContent component with Resizable sub-panel split.
  Both TerminalPane (focus mode) and CanvasTile (canvas mode) delegate
  terminal rendering to it. ~80% code deduplication.
- F1: Sub-panel derivations (hasSubs, isExpanded, activeSubTab, focusTarget)
  live in one place (TerminalContent) instead of being re-derived in two.
- F2: Focus coordination consolidated — TerminalContent internally pairs
  setFocusTarget with the external onFocus callback. Callers pass one prop.
- P3: handleSizesChange now collapses the sub-panel when dragged below
  threshold instead of silently discarding the resize.
- P1: Skipped — merging canvas layout with sub-panel state would complect
  independent concerns (spatial position vs split configuration).
